+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.3](https://github.com/silentium-lab/silentium-components/compare/v0.0.2...v0.0.3) (2025-04-16)
+
+
+### Features
+
+* **11-issue:** hash table structure ([dc9c588](https://github.com/silentium-lab/silentium-components/commit/dc9c588805f9205e728a6f52e01e6c4cc3769921))
+* **11-issue:** wip for hash table structure ([3dec97a](https://github.com/silentium-lab/silentium-components/commit/3dec97a78dbae97bf1f6ba1d7b5dfae4f900b919))
+
 ### [0.0.2](https://github.com/silentium-lab/silentium-components/compare/v0.0.1...v0.0.2) (2025-04-15)